src/requirements.txt
transformers
torch
accelerate>=0.26.0
openai==0.28
pinecone-client
pinecone
pandas
requests
psycopg2-binary
streamlit

